requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
pandas
streamlit>=1.32.0
matplotlib>=3.0
numpy-financial
scikit-learn>=1.3,<2.0
scipy>=1.10,<2.0
numpy
altair>=5.0
